pyproject.toml
[project]
name = "thales-cdsp-csm-mcp-server"
version = "1.0.0"
description = "Thales CDSP CSM Secrets Management MCP Server"
authors = [
    {name = "Sanyam Bassi", email = "sanyam.bassi@thalesgroup.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "requests>=2.25.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
thales-cdsp-csm-mcp-server = "src.thales_cdsp_csm_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "requests>=2.25.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

 